# Optimize light updates when turning conductors on and off